requirements.txt
accelerate
datasets
debugpy
huggingface-hub
ipykernel
ipython
kaggle
matplotlib
matplotlib-inline
multiprocess
numba
numpy
pandas
pillow
pymp
pymp-pypi
pyparsing
PyYAML
pyzmq
regex
safetensors
scikit-learn
scipy
seaborn
sympy
tiktoken
timm
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
triton
peft